split up osu! apis into separate files (clean up main.py)
also make a redis adapter
