+++ server/clients.py
@@ -3,10 +3,10 @@ from typing import TYPE_CHECKING
 import aiosu
 from aioredlock import Aioredlock
 from httpx import AsyncClient
-from redis.asyncio import Redis
 
 from server import settings
 from server.adapters.database import Database
+from server.adapters.redis import Redis
 
 if TYPE_CHECKING:
     from types_aiobotocore_s3.client import S3Client
